Refactor tfsec running
move all the code for running tfsec into its own class
