+++ src/explorer/utils.ts
@@ -1,7 +1,5 @@
 
-import * as vscode from 'vscode';
-import * as child from 'child_process';
-import { TfsecTreeItem } from './explorer/tfsec_treeitem';
+import { TfsecTreeItem } from './tfsec_treeitem';
 
 
 function getSeverityPosition(severity: string): number {
@@ -76,42 +74,7 @@ const uniqueLocations = (input: TfsecTreeItem[]): TfsecTreeItem[] => {
 };
 
 
-const getBinaryPath = () => {
-    const config = vscode.workspace.getConfiguration('tfsec');
-    var binary = config.get('binaryPath', 'tfsec');
-    if (binary === "") {
-        binary = "tfsec";
-    }
-
-    return binary;
-};
-
-const checkTfsecInstalled = (outputChannel: vscode.OutputChannel): boolean => {
-    const binaryPath = getBinaryPath();
-
-    var command = [];
-    command.push(binaryPath);
-    command.push('--help');
-    try {
-        child.execSync(command.join(' '));
-    }
-    catch (err) {
-        outputChannel.appendLine(`tfsec not found. Check the tfsec extension settings to ensure the path is correct. [${binaryPath}]`);
-        return false;
-    }
-    return true;
-};
-
-const getInstalledTfsecVersion = () => {
-    let binary = getBinaryPath();
-
-    var command = [];
-    command.push(binary);
-    command.push('--version');
-    const getVersion = child.execSync(command.join(' '));
-    return getVersion.toString();
-};
 
 const capitalize = (s: string) => (s && s[0] && s[0].toUpperCase() + s.slice(1).toLowerCase()) || '';
 
-export { getBinaryPath, sortByCode, sortBySeverity, uniqueLocations, getInstalledTfsecVersion, capitalize, checkTfsecInstalled };
+export { sortByCode, sortBySeverity, uniqueLocations, capitalize };
